exploring different ergonomics of interfacing go structs with SQL
looking into sqlx next, tempted to full send this home-brewed mini-ORM with Resource generics that can create tables and convert to JSON/HTMX/DOT with minimal repetitions but not trying to overdo it on the scaffolding with this project whose DB schema will not likely be changing at all... preparing for the wits-go interface to the normalized replays data, and the equivalent of ORM for GEL further out

+++ player.go
@@ -34,7 +34,7 @@ type Player struct {
 }
 
 // Simple (no GCID) constructor for a Player instance.
-func NewPlayer(id int, name string) Player {
+func NewPlayer(id int64, name string) Player {
 	return Player{ID: PlayerID{RowID: id}, Name: name}
 }
 
@@ -54,7 +54,7 @@ type PlayerRole struct {
 // Represents both (or either of) the internal identity and the player's GCID.
 // A zero value for either one will indicate missing or unknown value.
 type PlayerID struct {
-	RowID int
+	RowID int64
 	GCID  string
 }
